requirements.txt
# AlgoBot — Python dependencies
# Install: pip install -r requirements.txt

# Core
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Terminal display
rich>=13.0.0
click>=8.1.0

# Database
supabase>=2.0.0

# Market data
ccxt>=4.0.0
yfinance>=0.2.28
requests>=2.31.0
websocket-client>=1.6.0
aiohttp>=3.9.0
ib_insync>=0.9.86

# AI / ML
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0
lightgbm>=4.0.0
optuna>=3.6.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
schedule>=1.2.0
tabulate>=0.9.0
uuid>=1.30

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# MT5 broker (Windows only — comment out on Mac/Linux)
MetaTrader5>=5.0.45; sys_platform == "win32"
